+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+## [0.0.4] - 2022-09-11
+
+- Added command line options load and build
+- Added wait-until to rails w/postgresql or mysql to wait for the db to be ready before starting rails
+- Fixed per-state display not working
+- Use table_print to display some information such as linked projects
+
 ## [0.0.1] - 2022-06-26
 
 - Initial release